requirements.txt
Django>=5.0
djangorestframework>=3.14
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9
httpx>=0.25
python-dateutil>=2.8
markdown>=3.5
Pygments>=2.17
django-filter>=23
APScheduler>=3.10
